events: extract `on`, `once` and `getEventListeners` to a separate file
